+++ rosariosis.sql
@@ -490,6 +490,7 @@ CREATE TABLE accounting_incomes (
     id serial PRIMARY KEY,
     title text,
     amount numeric,
+    file_attached text,
     school_id integer NOT NULL,
     syear numeric(4,0) NOT NULL,
     created_at timestamp DEFAULT current_timestamp,