+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 8.4
 - Comments length > 60 chars, responsive table ColorBox in EditReportCardGrades.php, FinalGrades.php, Grades.php, InputFinalGrades.php & StudentGrades.php
 - Add File Attached to Incomes in Incomes.php & Accounting/functions.inc.php
 - Add File Attached to Expenses in Expenses.php
+- SQL ACCOUNTING_INCOMES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 
 Changes in 8.3.1
 ----------------